refactor: search bar component
